+++ src/graph/operators/sum.cpp
@@ -46,7 +46,7 @@ Node_handle sum(Computation_graph &rec, Node_handle A, int dim,
     INode *A_ptr = rec.get_node(A);
     Tensor &A_val = A_ptr->value();
 
-    if (dim < 0 || dim >= A_val.rank()) {
+    if (dim < 0 || dim >= static_cast<int>(A_val.rank())) {
         throw argument_error(
             make_graph_errmsg("argument error", recLen, "sum",
                               "dim has to be a valid index of A.shape",
